Feat&Fix: 拆分逻辑，增加若干功能，修改选种逻辑
