+++ app/flood_data.json
@@ -1,0 +1,56 @@
+[
+    {
+        "name": "AUDIOBOOK Palace struggle play Novels Collection69",
+        "id": "807079",
+        "time": "2024-08-02 11:08:39",
+        "size": 26517104220,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=1dc54d3befbde3c15f4c066c64be3d91&t=1722568113&tid=807079&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    },
+    {
+        "name": "AUDIOBOOK Urban life novels Collection158",
+        "id": "807071",
+        "time": "2024-08-02 11:09:01",
+        "size": 31213680064,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=7414857cfb5ac243c98ed9deaaf792af&t=1722568137&tid=807071&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    },
+    {
+        "name": "AUDIOBOOK YUNDUANCANYANG Collection3",
+        "id": "807068",
+        "time": "2024-08-02 11:09:24",
+        "size": 24051597696,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=56de1b143155137248b6ef47dbc03e39&t=1722568161&tid=807068&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    },
+    {
+        "name": "AUDIOBOOK Urban life novels Collection159",
+        "id": "807073",
+        "time": "2024-08-02 11:09:45",
+        "size": 30924961440,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=3dd5ffaac3b8fe15afbda1139788f0dd&t=1722568181&tid=807073&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    },
+    {
+        "name": "AUDIOBOOK GUBEI Collection1",
+        "id": "807094",
+        "time": "2024-08-02 11:10:00",
+        "size": 25870222882,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=091fb4f79614c299778b5f3f9c5cc54f&t=1722568197&tid=807094&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    },
+    {
+        "name": "AUDIOBOOK RUOLI Collection2",
+        "id": "807089",
+        "time": "2024-08-02 11:10:19",
+        "size": 19625130553,
+        "url": "https://api.m-team.cc/api/rss/dlv2?useHttps=true&type=ipv6&sign=78cdce76ae40222b1a95da014169e544&t=1722568217&tid=807089&uid=305724",
+        "discount": "FREE",
+        "discount_end_time": null
+    }
+]
