+++ app/requirements.txt
@@ -1,5 +1,2 @@
-certifi==2024.2.2
-charset-normalizer==3.3.2
-idna==3.6
-requests==2.31.0
-urllib3==2.2.1
+requests
+python-dotenv
